+++ tests/test_hyp3.py
@@ -1,7 +1,6 @@
 from datetime import datetime, timedelta
 from urllib.parse import urljoin
 
-import pytest
 import responses
 
 import hyp3_sdk
@@ -42,8 +41,7 @@ def test_find_jobs(get_mock_job):
     batch = api.find_jobs()
     assert len(batch) == 3
 
-    with pytest.warns(UserWarning):
-        batch = api.find_jobs()
+    batch = api.find_jobs()
     assert len(batch) == 0